Move all editor format and marker definitions to EditorFeatures.Wpf
These speak in terms of WPF colors and brushes, so must be moved out.
