+++ src/EditorFeatures/Core.Wpf/ConflictTagDefinition.cs
