+++ src/EditorFeatures/Core.Wpf/RenameTracking/RenameTrackingTagDefinition.cs
