+++ src/EditorFeatures/Core/Implementation/BraceMatching/ClassificationTypeDefinitions.cs
@@ -1,7 +1,6 @@
 ﻿// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
 
 using System.ComponentModel.Composition;
-using System.Windows.Media;
 using Microsoft.VisualStudio.Language.StandardClassification;
 using Microsoft.VisualStudio.Text.Classification;
 using Microsoft.VisualStudio.Utilities;
@@ -16,19 +15,5 @@ namespace Microsoft.CodeAnalysis.Editor.Implementation.BraceMatching
         [Name(BraceMatchingName)]
         [BaseDefinition(PredefinedClassificationTypeNames.FormalLanguage)]
         internal readonly ClassificationTypeDefinition BraceMatching;
-
-        [Export(typeof(EditorFormatDefinition))]
-        [ClassificationType(ClassificationTypeNames = BraceMatchingName)]
-        [Name(BraceMatchingName)]
-        [Order(After = Priority.High)]
-        [UserVisible(true)]
-        private class BraceMatchingFormatDefinition : ClassificationFormatDefinition
-        {
-            private BraceMatchingFormatDefinition()
-            {
-                this.DisplayName = EditorFeaturesResources.Brace_Matching;
-                this.BackgroundColor = Color.FromRgb(0xDB, 0xE0, 0xCC);
-            }
-        }
     }
 }
