+++ src/EditorFeatures/Core/Implementation/Classification/ClassificationTypeDefinitions.cs
@@ -2,7 +2,6 @@
 
 using System.ComponentModel.Composition;
 using System.Diagnostics.CodeAnalysis;
-using System.Windows.Media;
 using Microsoft.CodeAnalysis.Classification;
 using Microsoft.VisualStudio.Language.StandardClassification;
 using Microsoft.VisualStudio.Text.Classification;
@@ -17,63 +16,18 @@ namespace Microsoft.CodeAnalysis.Editor.Implementation.Classification
         [Name(ClassificationTypeNames.PreprocessorText)]
         [BaseDefinition(PredefinedClassificationTypeNames.FormalLanguage)]
         internal ClassificationTypeDefinition PreprocessorTextTypeDefinition { get; set; }
-
-        [Export(typeof(EditorFormatDefinition))]
-        [ClassificationType(ClassificationTypeNames = ClassificationTypeNames.PreprocessorText)]
-        [Name(ClassificationTypeNames.PreprocessorText)]
-        [Order(After = LanguagePriority.NaturalLanguage, Before = LanguagePriority.FormalLanguage)]
-        [UserVisible(true)]
-        [ExcludeFromCodeCoverage]
-        private class PreprocessorTextFormatDefinition : ClassificationFormatDefinition
-        {
-            private PreprocessorTextFormatDefinition()
-            {
-                this.DisplayName = EditorFeaturesResources.Preprocessor_Text;
-                this.ForegroundColor = Colors.Black;
-            }
-        }
         #endregion
         #region Punctuation
         [Export]
         [Name(ClassificationTypeNames.Punctuation)]
         [BaseDefinition(PredefinedClassificationTypeNames.FormalLanguage)]
         internal ClassificationTypeDefinition PunctuationTypeDefinition;
-
-        [Export(typeof(EditorFormatDefinition))]
-        [ClassificationType(ClassificationTypeNames = ClassificationTypeNames.Punctuation)]
-        [Name(ClassificationTypeNames.Punctuation)]
-        [Order(After = LanguagePriority.NaturalLanguage, Before = LanguagePriority.FormalLanguage)]
-        [UserVisible(true)]
-        [ExcludeFromCodeCoverage]
-        private class PunctuationFormatDefinition : ClassificationFormatDefinition
-        {
-            private PunctuationFormatDefinition()
-            {
-                this.DisplayName = EditorFeaturesResources.Punctuation;
-                this.ForegroundColor = Colors.Black;
-            }
-        }
         #endregion
         #region String - Verbatim
         [Export]
         [Name(ClassificationTypeNames.VerbatimStringLiteral)]
         [BaseDefinition(PredefinedClassificationTypeNames.FormalLanguage)]
         internal readonly ClassificationTypeDefinition StringVerbatimTypeDefinition;
-
-        [Export(typeof(EditorFormatDefinition))]
-        [ClassificationType(ClassificationTypeNames = ClassificationTypeNames.VerbatimStringLiteral)]
-        [Name(ClassificationTypeNames.VerbatimStringLiteral)]
-        [Order(After = LanguagePriority.NaturalLanguage, Before = LanguagePriority.FormalLanguage)]
-        [UserVisible(true)]
-        [ExcludeFromCodeCoverage]
-        private class StringVerbatimFormatDefinition : ClassificationFormatDefinition
-        {
-            private StringVerbatimFormatDefinition()
-            {
-                this.DisplayName = EditorFeaturesResources.String_Verbatim;
-                this.ForegroundColor = Colors.Maroon;
-            }
-        }
         #endregion
 
         #region User Types - Classes
@@ -81,153 +35,42 @@ namespace Microsoft.CodeAnalysis.Editor.Implementation.Classification
         [Name(ClassificationTypeNames.ClassName)]
         [BaseDefinition(PredefinedClassificationTypeNames.FormalLanguage)]
         internal readonly ClassificationTypeDefinition UserTypeClassesTypeDefinition;
-
-        [Export(typeof(EditorFormatDefinition))]
-        [ClassificationType(ClassificationTypeNames = ClassificationTypeNames.ClassName)]
-        [Name(ClassificationTypeNames.ClassName)]
-        [Order(After = PredefinedClassificationTypeNames.Identifier)]
-        [Order(After = PredefinedClassificationTypeNames.Keyword)]
-        [UserVisible(true)]
-        [ExcludeFromCodeCoverage]
-        private class UserTypeClassesFormatDefinition : ClassificationFormatDefinition
-        {
-            private UserTypeClassesFormatDefinition()
-            {
-                this.DisplayName = EditorFeaturesResources.User_Types_Classes;
-                this.ForegroundColor = Color.FromRgb(0x2B, 0x91, 0xAF);
-            }
-        }
         #endregion
         #region User Types - Delegates 
         [Export]
         [Name(ClassificationTypeNames.DelegateName)]
         [BaseDefinition(PredefinedClassificationTypeNames.FormalLanguage)]
         internal readonly ClassificationTypeDefinition UserTypeDelegatesTypeDefinition;
-
-        [Export(typeof(EditorFormatDefinition))]
-        [ClassificationType(ClassificationTypeNames = ClassificationTypeNames.DelegateName)]
-        [Name(ClassificationTypeNames.DelegateName)]
-        [Order(After = PredefinedClassificationTypeNames.Identifier)]
-        [Order(After = PredefinedClassificationTypeNames.Keyword)]
-        [UserVisible(true)]
-        [ExcludeFromCodeCoverage]
-        private class UserTypeDelegatesFormatDefinition : ClassificationFormatDefinition
-        {
-            private UserTypeDelegatesFormatDefinition()
-            {
-                this.DisplayName = EditorFeaturesResources.User_Types_Delegates;
-                this.ForegroundColor = Color.FromRgb(0x2B, 0x91, 0xAF);
-            }
-        }
         #endregion
         #region User Types - Enums 
         [Export]
         [Name(ClassificationTypeNames.EnumName)]
         [BaseDefinition(PredefinedClassificationTypeNames.FormalLanguage)]
         internal readonly ClassificationTypeDefinition UserTypeEnumsTypeDefinition;
-
-        [Export(typeof(EditorFormatDefinition))]
-        [ClassificationType(ClassificationTypeNames = ClassificationTypeNames.EnumName)]
-        [Name(ClassificationTypeNames.EnumName)]
-        [Order(After = PredefinedClassificationTypeNames.Identifier)]
-        [Order(After = PredefinedClassificationTypeNames.Keyword)]
-        [UserVisible(true)]
-        [ExcludeFromCodeCoverage]
-        private class UserTypeEnumsFormatDefinition : ClassificationFormatDefinition
-        {
-            private UserTypeEnumsFormatDefinition()
-            {
-                this.DisplayName = EditorFeaturesResources.User_Types_Enums;
-                this.ForegroundColor = Color.FromRgb(0x2B, 0x91, 0xAF);
-            }
-        }
         #endregion
         #region User Types - Interfaces 
         [Export]
         [Name(ClassificationTypeNames.InterfaceName)]
         [BaseDefinition(PredefinedClassificationTypeNames.FormalLanguage)]
         internal readonly ClassificationTypeDefinition UserTypeInterfacesTypeDefinition;
-
-        [Export(typeof(EditorFormatDefinition))]
-        [ClassificationType(ClassificationTypeNames = ClassificationTypeNames.InterfaceName)]
-        [Name(ClassificationTypeNames.InterfaceName)]
-        [Order(After = PredefinedClassificationTypeNames.Identifier)]
-        [Order(After = PredefinedClassificationTypeNames.Keyword)]
-        [UserVisible(true)]
-        [ExcludeFromCodeCoverage]
-        private class UserTypeInterfacesFormatDefinition : ClassificationFormatDefinition
-        {
-            private UserTypeInterfacesFormatDefinition()
-            {
-                this.DisplayName = EditorFeaturesResources.User_Types_Interfaces;
-                this.ForegroundColor = Color.FromRgb(0x2B, 0x91, 0xAF);
-            }
-        }
         #endregion
         #region User Types - Modules 
         [Export]
         [Name(ClassificationTypeNames.ModuleName)]
         [BaseDefinition(PredefinedClassificationTypeNames.FormalLanguage)]
         internal readonly ClassificationTypeDefinition UserTypeModulesTypeDefinition;
-
-        [Export(typeof(EditorFormatDefinition))]
-        [ClassificationType(ClassificationTypeNames = ClassificationTypeNames.ModuleName)]
-        [Name(ClassificationTypeNames.ModuleName)]
-        [Order(After = PredefinedClassificationTypeNames.Identifier)]
-        [Order(After = PredefinedClassificationTypeNames.Keyword)]
-        [UserVisible(true)]
-        private class UserTypeModulesFormatDefinition : ClassificationFormatDefinition
-        {
-            private UserTypeModulesFormatDefinition()
-            {
-                this.DisplayName = EditorFeaturesResources.User_Types_Modules;
-                this.ForegroundColor = Color.FromRgb(43, 145, 175);
-            }
-        }
         #endregion
         #region User Types - Structures 
         [Export]
         [Name(ClassificationTypeNames.StructName)]
         [BaseDefinition(PredefinedClassificationTypeNames.FormalLanguage)]
         internal readonly ClassificationTypeDefinition UserTypeStructuresTypeDefinition;
-
-        [Export(typeof(EditorFormatDefinition))]
-        [ClassificationType(ClassificationTypeNames = ClassificationTypeNames.StructName)]
-        [Name(ClassificationTypeNames.StructName)]
-        [Order(After = PredefinedClassificationTypeNames.Identifier)]
-        [Order(After = PredefinedClassificationTypeNames.Keyword)]
-        [UserVisible(true)]
-        [ExcludeFromCodeCoverage]
-        private class UserTypeStructuresFormatDefinition : ClassificationFormatDefinition
-        {
-            private UserTypeStructuresFormatDefinition()
-            {
-                this.DisplayName = EditorFeaturesResources.User_Types_Structures;
-                this.ForegroundColor = Color.FromRgb(0x2B, 0x91, 0xAF);
-            }
-        }
         #endregion
         #region User Types - Type Parameters 
         [Export]
         [Name(ClassificationTypeNames.TypeParameterName)]
         [BaseDefinition(PredefinedClassificationTypeNames.FormalLanguage)]
         internal readonly ClassificationTypeDefinition UserTypeTypeParametersTypeDefinition;
-
-        [Export(typeof(EditorFormatDefinition))]
-        [ClassificationType(ClassificationTypeNames = ClassificationTypeNames.TypeParameterName)]
-        [Name(ClassificationTypeNames.TypeParameterName)]
-        [Order(After = PredefinedClassificationTypeNames.Identifier)]
-        [Order(After = PredefinedClassificationTypeNames.Keyword)]
-        [UserVisible(true)]
-        [ExcludeFromCodeCoverage]
-        private class UserTypeTypeParametersFormatDefinition : ClassificationFormatDefinition
-        {
-            private UserTypeTypeParametersFormatDefinition()
-            {
-                this.DisplayName = EditorFeaturesResources.User_Types_Type_Parameters;
-                this.ForegroundColor = Color.FromRgb(0x2B, 0x91, 0xAF);
-            }
-        }
         #endregion
 
         #region XML Doc Comments - Attribute Name 
@@ -235,212 +78,60 @@ namespace Microsoft.CodeAnalysis.Editor.Implementation.Classification
         [Name(ClassificationTypeNames.XmlDocCommentAttributeName)]
         [BaseDefinition(PredefinedClassificationTypeNames.FormalLanguage)]
         internal readonly ClassificationTypeDefinition XmlDocCommentAttributeNameTypeDefinition;
-
-        [Export(typeof(EditorFormatDefinition))]
-        [ClassificationType(ClassificationTypeNames = ClassificationTypeNames.XmlDocCommentAttributeName)]
-        [Name(ClassificationTypeNames.XmlDocCommentAttributeName)]
-        [Order(After = Priority.Default, Before = Priority.High)]
-        [UserVisible(true)]
-        [ExcludeFromCodeCoverage]
-        private class XmlDocCommentAttributeNameFormatDefinition : ClassificationFormatDefinition
-        {
-            private XmlDocCommentAttributeNameFormatDefinition()
-            {
-                this.DisplayName = EditorFeaturesResources.XML_Doc_Comments_Attribute_Name;
-                this.ForegroundColor = Color.FromRgb(0x80, 0x80, 0x80); // CIDARKGRAY
-            }
-        }
         #endregion
         #region XML Doc Comments - Attribute Quotes 
         [Export]
         [Name(ClassificationTypeNames.XmlDocCommentAttributeQuotes)]
         [BaseDefinition(PredefinedClassificationTypeNames.FormalLanguage)]
         internal readonly ClassificationTypeDefinition XmlDocCommentAttributeQuotesTypeDefinition;
-
-        [Export(typeof(EditorFormatDefinition))]
-        [ClassificationType(ClassificationTypeNames = ClassificationTypeNames.XmlDocCommentAttributeQuotes)]
-        [Name(ClassificationTypeNames.XmlDocCommentAttributeQuotes)]
-        [Order(After = Priority.Default, Before = Priority.High)]
-        [UserVisible(true)]
-        [ExcludeFromCodeCoverage]
-        private class XmlDocCommentAttributeQuotesFormatDefinition : ClassificationFormatDefinition
-        {
-            private XmlDocCommentAttributeQuotesFormatDefinition()
-            {
-                this.DisplayName = EditorFeaturesResources.XML_Doc_Comments_Attribute_Quotes;
-                this.ForegroundColor = Color.FromRgb(0x80, 0x80, 0x80); // CIDARKGRAY
-            }
-        }
         #endregion
         #region XML Doc Comments - Attribute Value 
         [Export]
         [Name(ClassificationTypeNames.XmlDocCommentAttributeValue)]
         [BaseDefinition(PredefinedClassificationTypeNames.FormalLanguage)]
         internal readonly ClassificationTypeDefinition XmlDocCommentAttributeValueTypeDefinition;
-
-        // definition of how format is represented in tools options.
-        // also specifies the default format.
-        [Export(typeof(EditorFormatDefinition))]
-        [ClassificationType(ClassificationTypeNames = ClassificationTypeNames.XmlDocCommentAttributeValue)]
-        [Name(ClassificationTypeNames.XmlDocCommentAttributeValue)]
-        [Order(After = Priority.Default, Before = Priority.High)]
-        [UserVisible(true)]
-        [ExcludeFromCodeCoverage]
-        private class XmlDocCommentAttributeValueFormatDefinition : ClassificationFormatDefinition
-        {
-            private XmlDocCommentAttributeValueFormatDefinition()
-            {
-                this.DisplayName = EditorFeaturesResources.XML_Doc_Comments_Attribute_Value;
-                this.ForegroundColor = Color.FromRgb(0x80, 0x80, 0x80); // CIDARKGRAY
-            }
-        }
         #endregion
         #region XML Doc Comments - CData Section 
         [Export]
         [Name(ClassificationTypeNames.XmlDocCommentCDataSection)]
         [BaseDefinition(PredefinedClassificationTypeNames.FormalLanguage)]
         internal readonly ClassificationTypeDefinition XmlDocCommentCDataSectionTypeDefinition;
-
-        [Export(typeof(EditorFormatDefinition))]
-        [ClassificationType(ClassificationTypeNames = ClassificationTypeNames.XmlDocCommentCDataSection)]
-        [Name(ClassificationTypeNames.XmlDocCommentCDataSection)]
-        [Order(After = LanguagePriority.NaturalLanguage, Before = LanguagePriority.FormalLanguage)]
-        [UserVisible(true)]
-        [ExcludeFromCodeCoverage]
-        private class XmlDocCommentCDataSectionFormatDefinition : ClassificationFormatDefinition
-        {
-            private XmlDocCommentCDataSectionFormatDefinition()
-            {
-                this.DisplayName = EditorFeaturesResources.XML_Doc_Comments_CData_Section;
-                this.ForegroundColor = Color.FromRgb(0x80, 0x80, 0x80);    // CIDARKGRAY    
-            }
-        }
         #endregion
         #region XML Doc Comments - Comment 
         [Export]
         [Name(ClassificationTypeNames.XmlDocCommentComment)]
         [BaseDefinition(PredefinedClassificationTypeNames.FormalLanguage)]
         internal readonly ClassificationTypeDefinition XmlDocCommentCommentTypeDefinition;
-
-        [Export(typeof(EditorFormatDefinition))]
-        [ClassificationType(ClassificationTypeNames = ClassificationTypeNames.XmlDocCommentComment)]
-        [Name(ClassificationTypeNames.XmlDocCommentComment)]
-        [Order(After = LanguagePriority.NaturalLanguage, Before = LanguagePriority.FormalLanguage)]
-        [UserVisible(true)]
-        [ExcludeFromCodeCoverage]
-        private class XmlDocCommentCommentFormatDefinition : ClassificationFormatDefinition
-        {
-            private XmlDocCommentCommentFormatDefinition()
-            {
-                this.DisplayName = EditorFeaturesResources.XML_Doc_Comments_Comment;
-                this.ForegroundColor = Color.FromRgb(0x80, 0x80, 0x80);    // CIDARKGRAY
-            }
-        }
         #endregion
         #region XML Doc Comments - Delimiter 
         [Export]
         [Name(ClassificationTypeNames.XmlDocCommentDelimiter)]
         [BaseDefinition(PredefinedClassificationTypeNames.FormalLanguage)]
         internal readonly ClassificationTypeDefinition XmlDocCommentDelimiterTypeDefinition;
-
-        [Export(typeof(EditorFormatDefinition))]
-        [ClassificationType(ClassificationTypeNames = ClassificationTypeNames.XmlDocCommentDelimiter)]
-        [Name(ClassificationTypeNames.XmlDocCommentDelimiter)]
-        [Order(After = LanguagePriority.NaturalLanguage, Before = LanguagePriority.FormalLanguage)]
-        [UserVisible(true)]
-        [ExcludeFromCodeCoverage]
-        private class XmlDocCommentDelimiterFormatDefinition : ClassificationFormatDefinition
-        {
-            private XmlDocCommentDelimiterFormatDefinition()
-            {
-                this.DisplayName = EditorFeaturesResources.XML_Doc_Comments_Delimiter;
-                this.ForegroundColor = Color.FromRgb(0x80, 0x80, 0x80); // CIDARKGRAY
-            }
-        }
         #endregion
         #region XML Doc Comments - Entity Reference
         [Export]
         [Name(ClassificationTypeNames.XmlDocCommentEntityReference)]
         [BaseDefinition(PredefinedClassificationTypeNames.FormalLanguage)]
         internal readonly ClassificationTypeDefinition XmlDocCommentEntityReferenceTypeDefinition;
-
-        [Export(typeof(EditorFormatDefinition))]
-        [ClassificationType(ClassificationTypeNames = ClassificationTypeNames.XmlDocCommentEntityReference)]
-        [Name(ClassificationTypeNames.XmlDocCommentEntityReference)]
-        [Order(After = LanguagePriority.NaturalLanguage, Before = LanguagePriority.FormalLanguage)]
-        [UserVisible(true)]
-        [ExcludeFromCodeCoverage]
-        private class XmlDocCommentEntityReferenceFormatDefinition : ClassificationFormatDefinition
-        {
-            private XmlDocCommentEntityReferenceFormatDefinition()
-            {
-                this.DisplayName = EditorFeaturesResources.XML_Doc_Comments_Entity_Reference;
-                this.ForegroundColor = Colors.Green;
-            }
-        }
         #endregion
         #region XML Doc Comments - Name
         [Export]
         [Name(ClassificationTypeNames.XmlDocCommentName)]
         [BaseDefinition(PredefinedClassificationTypeNames.FormalLanguage)]
         internal readonly ClassificationTypeDefinition XmlDocCommentNameTypeDefinition;
-
-        [Export(typeof(EditorFormatDefinition))]
-        [ClassificationType(ClassificationTypeNames = ClassificationTypeNames.XmlDocCommentName)]
-        [Name(ClassificationTypeNames.XmlDocCommentName)]
-        [Order(After = LanguagePriority.NaturalLanguage, Before = LanguagePriority.FormalLanguage)]
-        [UserVisible(true)]
-        [ExcludeFromCodeCoverage]
-        private class XmlDocCommentNameFormatDefinition : ClassificationFormatDefinition
-        {
-            private XmlDocCommentNameFormatDefinition()
-            {
-                this.DisplayName = EditorFeaturesResources.XML_Doc_Comments_Name;
-                this.ForegroundColor = Color.FromRgb(0x80, 0x80, 0x80); // CIDARKGRAY
-            }
-        }
         #endregion
         #region XML Doc Comments - Processing Instruction
         [Export]
         [Name(ClassificationTypeNames.XmlDocCommentProcessingInstruction)]
         [BaseDefinition(PredefinedClassificationTypeNames.FormalLanguage)]
         internal readonly ClassificationTypeDefinition XmlDocCommentProcessingInstructionTypeDefinition;
-
-        [Export(typeof(EditorFormatDefinition))]
-        [ClassificationType(ClassificationTypeNames = ClassificationTypeNames.XmlDocCommentProcessingInstruction)]
-        [Name(ClassificationTypeNames.XmlDocCommentProcessingInstruction)]
-        [Order(After = LanguagePriority.NaturalLanguage, Before = LanguagePriority.FormalLanguage)]
-        [UserVisible(true)]
-        [ExcludeFromCodeCoverage]
-        private class XmlDocCommentProcessingInstructionFormatDefinition : ClassificationFormatDefinition
-        {
-            private XmlDocCommentProcessingInstructionFormatDefinition()
-            {
-                this.DisplayName = EditorFeaturesResources.XML_Doc_Comments_Processing_Instruction;
-                this.ForegroundColor = Color.FromRgb(0x80, 0x80, 0x80); // CIDARKGRAY
-            }
-        }
         #endregion
         #region XML Doc Comments - Text 
         [Export]
         [Name(ClassificationTypeNames.XmlDocCommentText)]
         [BaseDefinition(PredefinedClassificationTypeNames.FormalLanguage)]
         internal readonly ClassificationTypeDefinition XmlDocCommentTextTypeDefinition;
-
-        [Export(typeof(EditorFormatDefinition))]
-        [ClassificationType(ClassificationTypeNames = ClassificationTypeNames.XmlDocCommentText)]
-        [Name(ClassificationTypeNames.XmlDocCommentText)]
-        [Order(After = Priority.Default, Before = Priority.High)]
-        [UserVisible(true)]
-        [ExcludeFromCodeCoverage]
-        private class XmlDocCommentTextFormatDefinition : ClassificationFormatDefinition
-        {
-            private XmlDocCommentTextFormatDefinition()
-            {
-                this.DisplayName = EditorFeaturesResources.XML_Doc_Comments_Text;
-                this.ForegroundColor = Colors.Green;
-            }
-        }
         #endregion
 
         #region VB XML Literals - Attribute Name 
@@ -448,221 +139,66 @@ namespace Microsoft.CodeAnalysis.Editor.Implementation.Classification
         [Name(ClassificationTypeNames.XmlLiteralAttributeName)]
         [BaseDefinition(PredefinedClassificationTypeNames.FormalLanguage)]
         internal readonly ClassificationTypeDefinition XmlLiteralAttributeNameTypeDefinition;
-
-        [Export(typeof(EditorFormatDefinition))]
-        [ClassificationType(ClassificationTypeNames = ClassificationTypeNames.XmlLiteralAttributeName)]
-        [Name(ClassificationTypeNames.XmlLiteralAttributeName)]
-        [Order(After = LanguagePriority.NaturalLanguage, Before = LanguagePriority.FormalLanguage)]
-        [UserVisible(true)]
-        private class XmlLiteralAttributeNameFormatDefinition : ClassificationFormatDefinition
-        {
-            private XmlLiteralAttributeNameFormatDefinition()
-            {
-                this.DisplayName = EditorFeaturesResources.VB_XML_Literals_Attribute_Name;
-                this.ForegroundColor = Color.FromRgb(185, 100, 100); // HC_LIGHTRED
-            }
-        }
         #endregion
         #region VB XML Literals - Attribute Quotes 
         [Export]
         [Name(ClassificationTypeNames.XmlLiteralAttributeQuotes)]
         [BaseDefinition(PredefinedClassificationTypeNames.FormalLanguage)]
         internal readonly ClassificationTypeDefinition XmlLiteralAttributeQuotesTypeDefinition;
-
-        [Export(typeof(EditorFormatDefinition))]
-        [ClassificationType(ClassificationTypeNames = ClassificationTypeNames.XmlLiteralAttributeQuotes)]
-        [Name(ClassificationTypeNames.XmlLiteralAttributeQuotes)]
-        [Order(After = LanguagePriority.NaturalLanguage, Before = LanguagePriority.FormalLanguage)]
-        [UserVisible(true)]
-        private class XmlLiteralAttributeQuotesFormatDefinition : ClassificationFormatDefinition
-        {
-            private XmlLiteralAttributeQuotesFormatDefinition()
-            {
-                this.DisplayName = EditorFeaturesResources.VB_XML_Literals_Attribute_Quotes;
-                this.ForegroundColor = Color.FromRgb(85, 85, 85); // HC_LIGHTBLACK
-            }
-        }
         #endregion
         #region VB XML Literals - Attribute Value 
         [Export]
         [Name(ClassificationTypeNames.XmlLiteralAttributeValue)]
         [BaseDefinition(PredefinedClassificationTypeNames.FormalLanguage)]
         internal readonly ClassificationTypeDefinition XmlLiteralAttributeValueTypeDefinition;
-
-        [Export(typeof(EditorFormatDefinition))]
-        [ClassificationType(ClassificationTypeNames = ClassificationTypeNames.XmlLiteralAttributeValue)]
-        [Name(ClassificationTypeNames.XmlLiteralAttributeValue)]
-        [Order(After = LanguagePriority.NaturalLanguage, Before = LanguagePriority.FormalLanguage)]
-        [UserVisible(true)]
-        private class XmlLiteralAttributeValueFormatDefinition : ClassificationFormatDefinition
-        {
-            private XmlLiteralAttributeValueFormatDefinition()
-            {
-                this.DisplayName = EditorFeaturesResources.VB_XML_Literals_Attribute_Value;
-                this.ForegroundColor = Color.FromRgb(100, 100, 185); // HC_LIGHTBLUE
-            }
-        }
         #endregion
         #region VB XML Literals - CData Section 
         [Export]
         [Name(ClassificationTypeNames.XmlLiteralCDataSection)]
         [BaseDefinition(PredefinedClassificationTypeNames.FormalLanguage)]
         internal readonly ClassificationTypeDefinition XmlLiteralCDataSectionTypeDefinition;
-
-        [Export(typeof(EditorFormatDefinition))]
-        [ClassificationType(ClassificationTypeNames = ClassificationTypeNames.XmlLiteralCDataSection)]
-        [Name(ClassificationTypeNames.XmlLiteralCDataSection)]
-        [Order(After = LanguagePriority.NaturalLanguage, Before = LanguagePriority.FormalLanguage)]
-        [UserVisible(true)]
-        private class XmlLiteralCDataSectionFormatDefinition : ClassificationFormatDefinition
-        {
-            private XmlLiteralCDataSectionFormatDefinition()
-            {
-                this.DisplayName = EditorFeaturesResources.VB_XML_Literals_CData_Section;
-                this.ForegroundColor = Color.FromRgb(192, 192, 192); // HC_LIGHTGRAY
-            }
-        }
         #endregion
         #region VB XML Literals - Comment 
         [Export]
         [Name(ClassificationTypeNames.XmlLiteralComment)]
         [BaseDefinition(PredefinedClassificationTypeNames.FormalLanguage)]
         internal readonly ClassificationTypeDefinition XmlLiteralCommentTypeDefinition;
-
-        [Export(typeof(EditorFormatDefinition))]
-        [ClassificationType(ClassificationTypeNames = ClassificationTypeNames.XmlLiteralComment)]
-        [Name(ClassificationTypeNames.XmlLiteralComment)]
-        [Order(After = LanguagePriority.NaturalLanguage, Before = LanguagePriority.FormalLanguage)]
-        [UserVisible(true)]
-        private class XmlLiteralCommentFormatDefinition : ClassificationFormatDefinition
-        {
-            private XmlLiteralCommentFormatDefinition()
-            {
-                this.DisplayName = EditorFeaturesResources.VB_XML_Literals_Comment;
-                this.ForegroundColor = Color.FromRgb(98, 151, 85); // HC_LIGHTGREEN
-            }
-        }
         #endregion
         #region VB XML Literals - Delimiter 
         [Export]
         [Name(ClassificationTypeNames.XmlLiteralDelimiter)]
         [BaseDefinition(PredefinedClassificationTypeNames.FormalLanguage)]
         internal readonly ClassificationTypeDefinition XmlLiteralDelimiterTypeDefinition;
-
-        [Export(typeof(EditorFormatDefinition))]
-        [ClassificationType(ClassificationTypeNames = ClassificationTypeNames.XmlLiteralDelimiter)]
-        [Name(ClassificationTypeNames.XmlLiteralDelimiter)]
-        [Order(After = LanguagePriority.NaturalLanguage, Before = LanguagePriority.FormalLanguage)]
-        [UserVisible(true)]
-        private class XmlLiteralDelimiterFormatDefinition : ClassificationFormatDefinition
-        {
-            private XmlLiteralDelimiterFormatDefinition()
-            {
-                this.DisplayName = EditorFeaturesResources.VB_XML_Literals_Delimiter;
-                this.ForegroundColor = Color.FromRgb(100, 100, 185); // HC_LIGHTBLUE
-            }
-        }
         #endregion
         #region VB XML Literals - Embedded Expression 
         [Export]
         [Name(ClassificationTypeNames.XmlLiteralEmbeddedExpression)]
         [BaseDefinition(PredefinedClassificationTypeNames.FormalLanguage)]
         internal readonly ClassificationTypeDefinition XmlLiteralEmbeddedExpressionTypeDefinition;
-
-        [Export(typeof(EditorFormatDefinition))]
-        [ClassificationType(ClassificationTypeNames = ClassificationTypeNames.XmlLiteralEmbeddedExpression)]
-        [Name(ClassificationTypeNames.XmlLiteralEmbeddedExpression)]
-        [Order(After = LanguagePriority.NaturalLanguage, Before = LanguagePriority.FormalLanguage)]
-        [UserVisible(true)]
-        private class XmlLiteralEmbeddedExpressionFormatDefinition : ClassificationFormatDefinition
-        {
-            private XmlLiteralEmbeddedExpressionFormatDefinition()
-            {
-                this.DisplayName = EditorFeaturesResources.VB_XML_Literals_Embedded_Expression;
-                this.ForegroundColor = Color.FromRgb(85, 85, 85); // HC_LIGHTBLACK
-                this.BackgroundColor = Color.FromRgb(255, 254, 191); // HC_LIGHTYELLOW
-            }
-        }
         #endregion
         #region VB XML Literals - Entity Reference 
         [Export]
         [Name(ClassificationTypeNames.XmlLiteralEntityReference)]
         [BaseDefinition(PredefinedClassificationTypeNames.FormalLanguage)]
         internal readonly ClassificationTypeDefinition XmlLiteralEntityReferenceTypeDefinition;
-
-        [Export(typeof(EditorFormatDefinition))]
-        [ClassificationType(ClassificationTypeNames = ClassificationTypeNames.XmlLiteralEntityReference)]
-        [Name(ClassificationTypeNames.XmlLiteralEntityReference)]
-        [Order(After = LanguagePriority.NaturalLanguage, Before = LanguagePriority.FormalLanguage)]
-        [UserVisible(true)]
-        private class XmlLiteralEntityReferenceFormatDefinition : ClassificationFormatDefinition
-        {
-            private XmlLiteralEntityReferenceFormatDefinition()
-            {
-                this.DisplayName = EditorFeaturesResources.VB_XML_Literals_Entity_Reference;
-                this.ForegroundColor = Color.FromRgb(185, 100, 100); // HC_LIGHTRED
-            }
-        }
         #endregion
         #region VB XML Literals - Name 
         [Export]
         [Name(ClassificationTypeNames.XmlLiteralName)]
         [BaseDefinition(PredefinedClassificationTypeNames.FormalLanguage)]
         internal readonly ClassificationTypeDefinition XmlLiteralNameTypeDefinition;
-
-        [Export(typeof(EditorFormatDefinition))]
-        [ClassificationType(ClassificationTypeNames = ClassificationTypeNames.XmlLiteralName)]
-        [Name(ClassificationTypeNames.XmlLiteralName)]
-        [Order(After = LanguagePriority.NaturalLanguage, Before = LanguagePriority.FormalLanguage)]
-        [UserVisible(true)]
-        private class XmlLiteralNameFormatDefinition : ClassificationFormatDefinition
-        {
-            private XmlLiteralNameFormatDefinition()
-            {
-                this.DisplayName = EditorFeaturesResources.VB_XML_Literals_Name;
-                this.ForegroundColor = Color.FromRgb(132, 70, 70); // HC_LIGHTMAROON
-            }
-        }
         #endregion
         #region VB XML Literals - Processing Instruction 
         [Export]
         [Name(ClassificationTypeNames.XmlLiteralProcessingInstruction)]
         [BaseDefinition(PredefinedClassificationTypeNames.FormalLanguage)]
         internal readonly ClassificationTypeDefinition XmlLiteralProcessingInstructionTypeDefinition;
-
-        [Export(typeof(EditorFormatDefinition))]
-        [ClassificationType(ClassificationTypeNames = ClassificationTypeNames.XmlLiteralProcessingInstruction)]
-        [Name(ClassificationTypeNames.XmlLiteralProcessingInstruction)]
-        [Order(After = LanguagePriority.NaturalLanguage, Before = LanguagePriority.FormalLanguage)]
-        [UserVisible(true)]
-        private class XmlLiteralProcessingInstructionFormatDefinition : ClassificationFormatDefinition
-        {
-            private XmlLiteralProcessingInstructionFormatDefinition()
-            {
-                this.DisplayName = EditorFeaturesResources.VB_XML_Literals_Processing_Instruction;
-                this.ForegroundColor = Color.FromRgb(192, 192, 192); // HC_LIGHTGRAY
-            }
-        }
         #endregion
         #region VB XML Literals - Text 
         [Export]
         [Name(ClassificationTypeNames.XmlLiteralText)]
         [BaseDefinition(PredefinedClassificationTypeNames.FormalLanguage)]
         internal readonly ClassificationTypeDefinition XmlLiteralTextTypeDefinition;
-
-        [Export(typeof(EditorFormatDefinition))]
-        [ClassificationType(ClassificationTypeNames = ClassificationTypeNames.XmlLiteralText)]
-        [Name(ClassificationTypeNames.XmlLiteralText)]
-        [Order(After = LanguagePriority.NaturalLanguage, Before = LanguagePriority.FormalLanguage)]
-        [UserVisible(true)]
-        private class XmlLiteralTextFormatDefinition : ClassificationFormatDefinition
-        {
-            private XmlLiteralTextFormatDefinition()
-            {
-                this.DisplayName = EditorFeaturesResources.VB_XML_Literals_Text;
-                this.ForegroundColor = Color.FromRgb(85, 85, 85); // HC_LIGHTBLACK
-            }
-        }
         #endregion
     }
 }
